base: Remove now unnecessary 'with-store' replacement.
This 'with-store' macro has become unnecessary since Guix commit
8ed597f4a261fe188de82cd1f5daed83dba948eb, which changes its own
'with-store' macro to not rely on 'dynamic-wind'.

Furthermore, this definition of 'unwind-protect' would rewrap SRFI-34
exceptions on Guile 3.0 in something like this:

  #<&compound-exception components: (#<&error> #<&irritants irritants: (#<&evaluation-error name: "shepherd" id: 26>)> #<&exception-with-kind-and-args kind: %exception args: (#<&evaluation-error name: "shepherd" id: 26>)>)>

The 'error?' and 'evaluation-error?' predicates would return #f for
these things, thereby leaving an uncaught exception in the fiber calling
'process-specs'.

* src/cuirass/base.scm (with-store): Remove.
* src/cuirass/utils.scm (unwind-protect): Remove.