store: ‘build-derivations&’ enforces synchronization with the build.
Previously, it was possible for the user to call the returned thunk
before ‘build-derivations’ had completed, thereby getting #f (the
initial value of ‘result’).  This made no sense because
‘build-derivations’ always returns #t or raises an exception.

This situation could happen in ‘cuirass remote-worker’ if the
corresponding ‘cuirass remote-server’ process disappeared in the middle
of a build, because ‘send-logs’ would return early, leading the finish
thunk of ‘build-derivations&’ to be called before build completion.

This change uses a channel to enforce synchronization with the thread
that calls ‘build-derivations’.

Partly fixes <https://issues.guix.gnu.org/66692>.

* src/cuirass/store.scm (build-derivations&): Use
‘with-exception-handler’ and ‘put-message’ in the thread.  Use
‘get-message’ instead of ‘atomic-box-ref’ in the finalization procedure.

Reported-by: Maxim Cournoyer <maxim.cournoyer@gmail.com>